rename utils to assignment-ast
